switch api and models
