--- api/buf.gen.yaml
@@ -1,15 +1,0 @@
-version: v1
-managed:
-  enabled: true
-  go_package_prefix:
-    default: github.com/maxthom/mir/internal/ito
-plugins:
-  - plugin: go
-    out: ../internal/ito
-    opt:
-      - paths=source_relative
-  # - plugin: connect-go
-  #   out: gen
-  #   opt:
-  #     - paths=source_relative
-# TODO how to generate binary next to proto
