--- api/clients/health/alive.sh
@@ -1,1 +1,0 @@
-curl http://localhost:3000/alive
