--- api/clients/v1alpha/protoproxy/uploadschema.sh
@@ -1,3 +1,0 @@
-curl --header 'Content-Type: application/json' \
- --data '{"sentence": "I feel happy."}'\
- http://localhost:3000/api/v1alpha.protoproxy.ProtoProxyService/UploadSchema
