--- api/proto/v1alpha/core_ito/core.proto
@@ -1,182 +1,0 @@
-syntax = "proto3";
-
-package v1alpha.core;
-
-option go_package = "gen/";
-
-import "proto/v1alpha/common_ito/common.proto";
-
-/*
-  Send a request to create a new device in the system
-*/
-message CreateDeviceRequest {
-  string device_id = 1; // the device_id must be unique
-  string name = 2; // Name of the device
-  string namespace = 3; // Name of the device
-  bool disabled = 4; // If disabled, device will be cut off from communication
-  map<string, string> labels = 5; // labels are used to uniquely identify the device
-  map<string, string> annotations = 6; // annotations are used to add additional information
-}
-
-/*
-  Response to the create device request
-  Only the error or the response data will be set
-*/
-message CreateDeviceResponse {
-  oneof response {
-    DeviceList ok = 1; // list of device that were created
-    common.Error error = 2;  // HTTP like error if something went wrong
-  }
-}
-
-/*
-  Send a request to update a device in the system
-  - Optional field, means can be nil
-  - when nil, mean we do not need to set them
-  - when nil, for a map value, it means mark for deletion
-*/
-message UpdateDeviceRequest {
-  Targets targets = 1; // To target a set of devices, search using ids, labels and annotations
-
-  optional Meta meta = 2;
-  optional Spec spec = 3;
-  optional Status status = 4;
-
-  message Meta {
-    optional string name = 3; // Name of the device
-    optional string namespace = 4; // Name of the device
-    map<string, OptString> labels = 5; // set key to nil to remove it
-    map<string, OptString> annotations = 6; // set key to nil to remove it
-  }
-  message Spec {
-    optional bool disabled = 3; // If disabled, device will be cut off from communication
-  }
-  message Status {
-    optional bool online = 3; // Is device online ?
-    optional common.Timestamp last_hearthbeat = 4; // Last pulse received
-  }
-  message OptString {
-    optional string value = 1;
-  }
-}
-
-/*
-  Response to the create device request
-  Only the error or the response data will be set
-*/
-message UpdateDeviceResponse {
-  oneof response {
-    DeviceList ok = 1; //  List of device that were updated
-    common.Error error = 2; // HTTP like error if something went wrong
-  }
-}
-
-/*
-  Send a request to delete a device in the system
-*/
-message DeleteDeviceRequest {
-  Targets targets = 1; // To target a set of devices, search using ids, labels and annotations
-}
-
-/*
-  Response to the create device request
-  Only the error or the response data will be set
-*/
-message DeleteDeviceResponse {
-  oneof response {
-    DeviceList ok = 1; // List of device that were deleted
-    common.Error error = 2; // HTTP like error if something went wrong
-  }
-}
-
-/*
-  Send a request to list devices in the system
-*/
-message ListDeviceRequest {
-  Targets targets = 1; // To target a set of devices, search using ids, labels and annotations
-}
-
-/*
-  Response to the create device request
-  Only the error or the response data will be set
-*/
-message ListDeviceResponse {
-  oneof response {
-    DeviceList ok = 1; // Devices that were found from the search
-    common.Error error = 2;  // HTTP like error if something went wrong
-  }
-}
-
-/*
-  Targets is used to target a set of devices using
-  a combination of ids, labels and annotations
-*/
-message Targets {
-  repeated string ids = 1; // ids of the devices
-  repeated string names = 2; // names of the devices
-  repeated string namespaces = 3; // namespace of the devices
-  map<string, string> labels = 4; // Labels are used to uniquely identify the device
-  map<string, string> annotations = 5; // Annotations are used to add additional information
-}
-
-/*
-* A set of devices
-*/
-message DeviceList {
-  repeated Device devices = 1;
-}
-
-/*
-* A set of devices with only their Id
-*/
-message DeviceIdList {
-  repeated string device_ids = 1;
-}
-
-/*
-  Represent a device in the system
-*/
-message Device {
-  string id = 1; // Represent the database id of the device
-  string api_version = 2; // Current version of the spec
-  string api_name = 3; // Name of the spec
-  Meta meta = 4; // Represent the specifications that can be updated
-  Spec spec = 5; // Represent device twin specific config from Mir
-  Properties properties = 6; // Represent set of device defined properties
-  Status status = 7; // Represent the status reported by the Mir
-}
-
-/*
-* Set of data from the Mir system to describe a device
-* or an entity in the system
-*/
-message Meta {
-    string name = 1; // Name of the device
-    string namespace = 2; // Name of the device
-    map<string, string> labels = 3; // Labels are used to uniquely identify the device
-    map<string, string> annotations = 4; // Annotations are used to add additional information
-}
-
-/*
-* Represent configuration from Mir itself for the twin specifically
-*/
-message Spec {
-    string device_id = 1; // Represent the Mir id of the device
-    bool disabled = 2; // If disabled, device will be cut off from communication
-}
-
-/*
-* Devices desired properties from the users
-* and reported properties from the devices
-*/
-message Properties {
-
-}
-
-/*
-* Status fields describe the device conditions monitored by Mir
-*/
-message Status {
-  common.Timestamp last_hearthbeat = 8; // Represent the time when the last hearthbeat was received
-  bool online = 9; // The device is online and sending heathbeath signal
-}
