--- api/proto/v1alpha/device_ito/device.proto
@@ -1,44 +1,0 @@
-syntax = "proto3";
-
-package v1alpha.device;
-
-option go_package = "gen/";
-
-/*
-  Send a request to create a new device in the system
-*/
-message SchemaRetrieveRequest {
-}
-
-/*
-  Response to the create device request
-  Only the error or the response data will be set
-*/
-message SchemaRetrieveResponse {
-  oneof response {
-  	bytes schema = 1; // Schema in bytes
-    Error error = 2;  // HTTP like error if something went wrong
-  }
-}
-
-// TODO move to shared proto
-
-// Generic error message
-message Error {
-    uint32 code = 1;     // HTTP like status code
-    string message = 2;  // Human-readable error message.
-    repeated string details = 3; // Optional details.
-}
-
-message Timestamp {
-  // Represents seconds of UTC time since Unix epoch
-  // 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
-  // 9999-12-31T23:59:59Z inclusive.
-  int64 seconds = 1;
-
-  // Non-negative fractions of a second at nanosecond resolution. Negative
-  // second values with fractions must still have non-negative nanos values
-  // that count forward in time. Must be from 0 to 999,999,999
-  // inclusive.
-  int32 nanos = 2;
-}
