--- api/proto/v1alpha/protoflux_ito/protoflux.proto
@@ -1,16 +1,0 @@
-syntax = "proto3";
-
-package v1alpha.protoflux;
-
-option go_package = "gen/";
-
-message UploadSchemaRequest { }
-
-message UploadSchemaResponse {
-  string msg = 1;
-}
-
-service ProtoProxyService {
-  // have another store specific service
-  rpc UploadSchema(UploadSchemaRequest) returns (UploadSchemaResponse) {}
-}
