--- api/proto/v1alpha/protostore/protostore.proto
@@ -1,35 +1,0 @@
-syntax = "proto3";
-
-package v1alpha.protostore;
-
-option go_package = "gen/";
-
-message UploadRequest { }
-
-message UploadResponse {
-    string msg = 1;
-}
-
-message DownloadRequest { }
-
-message DownloadResponse {
-    string msg = 1;
-}
-
-message ListRequest { }
-
-message ListResponse {
-    string msg = 1;
-}
-
-
-service ProtoStoreService {
-// Upload a  to the store with attached metadata
-rpc Upload(UploadRequest) returns (UploadResponse) {}
-
-// Download the a  using the name or uuid
-rpc Download(DownloadRequest) returns (DownloadResponse) {}
-
-// List all  present in the store
-rpc List(ListRequest) returns (ListResponse) {}
-}
