+++ docs/TODO.md
@@ -1,0 +1,19 @@
+# TODO
+## Immediate
+
+- [ ] fix cycle of clients and mir module
+
+## Refactoring
+
+- [ ] create external with interfaces
+- [ ] hide itos from domain objects in sdk
+- [ ] look on how k8s sdk is made
+- [ ] ask redit for my questioning
+- [ ] need a centralized place for errors, in models
+- [ ] surreal api
+
+## Protoflux
+
+- [ ] integrate database
+- [ ] add unit test
+- [ ] work on protodash
