+++ examples/hearthbeat_device/main.go
@@ -15,7 +15,7 @@ func main() {
 	mir_signals.Notify(syscall.SIGHUP, syscall.SIGINT, syscall.SIGQUIT)
 
 	m, err := mir.Builder().
-		DeviceId("0xf86ea").
+		DeviceId("0xf86eb").
 		Target("nats://127.0.0.1:4222").
 		LogLevel(mir.LogLevelInfo).
 		LogWriters([]io.Writer{os.Stdout}).
