+++ pkgs/api/buf.gen.yaml
@@ -1,0 +1,15 @@
+version: v1
+managed:
+  enabled: true
+  go_package_prefix:
+    default: github.com/maxthom/mir/pkgs/api/proto
+plugins:
+  - plugin: go
+    out: proto
+    opt:
+      - paths=source_relative
+  # - plugin: connect-go
+  #   out: gen
+  #   opt:
+  #     - paths=source_relative
+# TODO how to generate binary next to proto
