+++ pkgs/module/mir/test/buf.gen.yaml
